requirements.txt
Flask==0.12.2
Flask-HTTPAuth==3.2.3
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
passlib==1.7.1
psycopg2==2.7.3.2